+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
+### Added
+
+- Added new menu item Diagnostics->Restart Application
+
 ### Fixed
 
 - Fixed [Pipeline] objects showing an ID of 0 in tree collections